+++ packages/react-server-webpack-plugin/src/templates/server.ts
@@ -1,15 +1,20 @@
+export default app;
+
 import React from 'react';
-import {createServer} from '@shopify/react-server';
+import {createServer, DefaultProvider} from '@shopify/react-server';
 import App from 'index';
 
 const render = ctx =>
-  React.createElement(App, {
-    server: true,
-    location: ctx.request.url,
-  });
+  React.createElement(
+    DefaultProvider,
+    null,
+    React.createElement(App, {
+      server: true,
+      location: ctx.request.url,
+    }),
+  );
 
 const app = createServer({
   render,
 });
-
 export default app;